Added redux store, actions, reducers and initialized redux in application
