+++ actions/types.js
@@ -1,0 +1,2 @@
+export const SET_LOCATION = "SET_LOCATION";
+export const SET_COVID_DATA = "SET_COVID_DATA";
